build: generate queries into conf/ directory
The queries are loaded during runtime and currently, they are
expected to be found in the schema/ directory. After the install
process is defined, the location will be different. As it is
treated as configuration file (ini), it is desirable to place
it directly under the ${config.dir}. For this purpose, we need
to output the queries ini file into the right place.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>