build: compile libraries in shared mode
This is an obvious step. The built binaries contain the RPATH however
the installed one do not. This is correct. Before installation, it is
possible to directly test the binaries without any further settings.
After installation is done, the libraries are search in the standard
way of the target system (looking for them in /usr/lib).

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>