TransactionalWorkRunner: new class
If the WorkExecutor requires a transaction for its execution, the TransactionalWorkRunner
should be used instead of the GenericWorkRunner. The execution is divided into two steps:

1. set work to Work::STATE_EXECUTED
2. execute the work and handle result

Both are potentially executed in a different transaction. This is due to the fact that
when the work is executed, we have to be sure that it state has been persisted. If
the application fails for some reason, we can find out that such work entered
the Work::STATE_EXECUTED state and avoid executing it again. It can be marked as
Work::STATE_FAILED after restart of the application.

If the WorkExecutor finishes the execution (either via suspend request, finish or
a failure) the work's state is changed appropriately. This must be done in the
some transaction the WorkExecutor has been running in. Otherwise, we can crash
inconsitently. E.g. the Work is marked as Work::STATE_EXECUTED, however, it has
been finished successfully.

Transactions cannot cover remote processing (like communication with a remote
gateway or remote sensors). Thus, some inconsistencies might happen anyway.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>