DeviceService: utilize DeviceUnpairWork
Implement DeviceService::unregister() via DeviceUnpairWork. The work
instance is created and scheduled for immediate execution. Currently,
there is no possible check whether another such work is in progress.
However, two concurrent executions for the same device should not
be in conflict here because both just call the GatewayRPC and then
polls the database.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>