PocoAbstractDao: handle bad_cast when creating session()
When a wrong TransactionManager is set and some Dao derived from
PocoAbstractDao asks for a session then bad_cast exception is
thrown. However, it does not contain any useful information and
makes searching the source of the issue diffcult. Convert the
bad_cast to a Poco Exception with a more sensible message.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>