PocoDaoManager: define and apply connectionTimeout property
Especially, when testing with SQLite with multiple sessions, we
have problem with accessing a locked database (due to the concurrent
access). To avoid this issue, it is possible to define busy timeout
(for SQLite [1]). This is set by calling [2]

  Poco::Data::Session::setConnectionTimeout(...)

This must be configured for each new session. Provide a new injectable
property connectionTimeout into PocoDaoManager and propagate this
property into PocoAbstractDao that applies it when creating a new
session, no matter whether it is transactional or not.

[1] https://sqlite.org/pragma.html#pragma_busy_timeout
[2] https://pocoproject.org/docs/Poco.Data.Session.html#7031

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>