t: xmlui: delay waiting for device unregister
When running tests against the implementation of unregister utilizing
the work-queue, the result is asynchronous and the test might fail
before the operation is done. This can be achieved e.g. by running
the server with valgrind.

Fix this by adding an approx 1 second delay when waiting for the
list of new devices. We ask until we get the right result.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>