UriPattern: new class
The UriPattern is an internal representation of URI for matching
requests to REST actions. This class is an internal implementation
that is invisible to the REST development API and can be reimplemented
in the future.

The UriPattern stores a pattern that can contain the following path
segments types:

* string (e.g. /gateways -> "gateways" is a string segment)
* wildcard (e.g. /*/devices -> "*" is a wildcard that matches anything)
* param (e.g. /devices/$deviceId -> "$deviceId" is a param that can be extracted)

There are also special patterns:

* "" (empty pattern) - does never match anything
* "*" - matches anything

Relative patterns does not match absolute path requests, e.g.:

* gateways/devices would never match /gateways/devices

(There is no usage for this behaviour and so it might be reimplemented
in the future.)

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>