RestAction: new class
The RestAction represents a REST action that is to be performed in the system
usually on a user request. The class is not part of the REST developer API
and can be changed or remove in the future. It holds information about the action
behaviour and othre requirements like whether the action must be executed with
a user session or what is the limit of the user input (security).

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>