JSONRestHandler: new class
The JSONRestHandler implements the BeeeOn application-specific JSON protocol.
It provides generic JSON procedure parseInput to parse any incoming JSON data
and application-specific routines to wrap a request result or to handle
exceptions.

The output is to be generated via Poco::JSON::PrintHandler. The input is
parsed into a Poco::JSON::Object instance.

The generated output is in two forms:

{
	"code": <HTTP Status>,
	"status": "success",
	"data": <Request-specific result>
}

of

{
	"code": <HTTP Status>,
	"status": {"error" | "fail"},
	"message": <String>
}

where the "error" status means that there is some application error (invalid
input) caused by the caller and "fail" status means that the server fails
to handle the request.

The client code should not generate errors on its own in most cases. The internal
logic of JSONRestHandler should generate appropriate erros based on the thrown
exceptions.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>