PocoContext: new class
The REST development API is based on the Poco::Net::HTTPServer and the related
objects. The PocoContext, PocoRequest and PocoResponse defines interfaces to
the request and response objects in a simple way while hiding the Poco::Net
specific details. This might simplify transition to another HTTP server if
necessary.

The PocoContext is just a template that provides a common methods for both
PocoRequest and PocoResponse and it is never used alone.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>