JSONGatewayDeserializer: new class
Implements deserialization of the Gateway according to the BeeeOn
Rest-UI API. It supports the following input formats:

1. GatewayRegister (full deserialization)

{
	"name": <String>,
	"altitude": <double|null>,
	"latitude": <double|null>,
	"longitude": <double|null>
}

Note that, the GatewayRegister format requires attribute "id". This
attribute must be parsed separately due to the GatewayService behaviour
(the Gateway instance must be created with a valid GatewayID.

2. GatewayUpdate (partial deserialization)

{
	"name": <String>,
	"altitude": <double|null>,
	"latitude": <double|null>,
	"longitude": <double|null>
}

The fields might be missing. The missing values are skipped (and thus
not updated).

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>