AuthRestHandler: new class
Implementation of the login/logout REST actions. It defines the following
REST actions:

1. auth.list   (should be GET)

Lists all available login possibilities (AuthProvider instances). Outputs
an array of JSON objects:

[
	{
		"id": <Provider Identification>,
		"name": <Display Name>,
		"oauth2": {
			"client_id": <Client ID>
		}
	},
	...
]

The "oauth2" is provided only for OAuth 2.0 AuthProvider instances.

2. auth.login  (should be POST)

Logins the user by providing a JSON object that is specific to the selected
provider. The expected format is:

{
	"provider": <Provider Identification>,
	"code": <OAuth 2.0 auth-code>
}

Outputs an object with the generated session ID:

{
	"id": <Session ID>
}

3. auth.logout (should be DELETE)

Logouts the current user. The input for this operation comes the from
the current session. Generates no output.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>