restui: define /auth URI mapping of AuthRestHandler
The BeeeOn RestUI API defines the following mapping to AuthRestHandler
actions:

1. GET /auth -> AuthRestHandler::list
2. POST /auth -> AuthRestHandler::login
3. DELETE /auth -> AuthRestHandler::logout

The new mapping scheme however breaks the test of unexisting path
via GET /auth in integration tests. As The GET /auth request is valid
now, we have to fix the test and call a different non-existing GET
path. This however changes the test's original purpose - to check
that an existing resource cannot be accessed via another method
then expected. Thus, we are adding test4 with PUT on /auth.

The integration tests are fixed to login correctly with this new
API scheme and JSON data format. The POST /auth returns a JSON object
with the session ID instead of just a plain string value, now.
The DELETE /auth now returns 204 instead of 200.