PocoRestRequestHandler: new class with factory
The PocoRestRequestHandler and PocoRestRequestFactory implements
the Poco::Net::HTTPRequestHandler and Poco::Net::HTTPRequestHandlerFactory
to integrate with Poco::Net::HTTPServer.

The PocoRestRequestFactory routes the incoming requests via the RestRouter
and creates the PocoRestRequestHandler handling the resolved RestAction.

There are two unmapped RestAction names defined that should be implemented
and handled by a application-specific RestHandler:

* builtin.unauthorized - the route being accessed requires a session however
  no session exists for the current HTTP request

* builtin.noroute - no route has been found, the current HTTP request is
  handled by this action, the params extracted via UriPattern are lost

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>