LegacyGatewayRPC: fix waiting for slow response
When receiving a response, it can come in very short byte amounts
sometimes. The current implementation would not properly read such
response. Fix this by waiting for the </reply> tag. Note that there
is probably no other way how to solve it because the communication
protocol does not provide any other metadata.

Fixes: 7d6aaaa2315c ("LegacyGatewayRPC: new class")
Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>
Reviewed-by: Jozef Halaj <xhalaj03@stud.fit.vutbr.cz>