Gateway: treat altitude as signed int
There is no reason to represent altitude as double. We are OK with
expressing in meters. Convert all altitude representations to signed int.

Also, the Android client expects the altitude as int, otherwise, it
does not parse it successfully. The home7 database representes altitude
as real.

The altitude can be missing - it is nullable. In the XML-UI protocol,
a null value is expressed as an empty string.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>