schema: introduce table devices_properties
Introduce a place where properties for certain device can be stored.
A generic content can be stored here. The properties are queried via
getparameters request from gateways.

Each property is specified as a pair (type, value). The type is an
integer specified by the application layer. The value is some text.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>