DeviceProperty: new class
Device property is some entry that extends information about a device.
Device properties are not stored with history (they are not temporal).
Only certain types of properties can be stored. The types are defined
by the enum DevicePropertyKey.

Some properties are stored in an encrypted way. This is to improve
security in case of a database data leakage. Passwords and IP addresses
of devices are protected this way.

The property firmware can be reported from device. It allows to report
a device firmware. Thus, it is read-only from the user UI.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>