XmlDevicePropertyDeserializer: new class
We can deserialize either for creation (full) or for update (partial).
Both methods uses however the same implementation because there is nothing
to be updated just partially.

In fact, we can consider to not change the CryptoParams on the partial
deserialization. However, we do it as it automatically updates the
record to contain a new salt on each update (probably more secure)
and in case of migration to another algorithm or other settings, it
automatically upgrades such encrypted property.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>