xmlui: serialize DecryptedDeviceProperty
The DeviceProperty itself cannot be serialized because it requires the Cipher
everytime. This complicates the design and methods implementation and introduces
dependencies on the cryptographical code into serialization. To avoid it, we
serialize only the DecryptedDeviceProperty instances.

Thus, anytime we would like to serialize DeviceProperty, it is necessary to
wrap it into a DecryptedDeviceProperty. Note that the property 'password' is
serialized by using placeholders ('*****').

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>