DecryptedDeviceProperty: new class
To simplify working with DeviceProperty instances holding encrypted data,
the DecryptedDeviceProperty can be used. Anywhere in the application when
we are safe to access the unencrypted data, we can wrap a DeviceProperty
into the DecryptedDeviceProperty and pass it on. The user of the wrapper
has full access to the property content.

The method DecryptedDeviceProperty::asPassword() returns a placeholder
as default instead of the actual password. Most of the time, the password
does not have to be decrypted and we only need to have some representation
for user.

The non-encrypted properties are accessible easily as well.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>