GoogleAuthProvider: validate provided token-related values
The access_token, token_type and expires_in are fields that must be
always provided. If they are missing, something went wrong.

The refresh_token is currently unused and according to the
documentation [1], it is present only in offline requests.
However, in the tests, the refresh_token is always present.
We ignore it for now.

Later, when access_type is used for other purposes then just
the login procedure and reading of user info, the refresh_token
might be stored for later use.

[1] https://developers.google.com/identity/protocols/OAuth2WebServer#handlingtheresponse

 refresh_token: A token that you can use to obtain a new access
                token. Refresh tokens are valid until the user
		revokes access. Again, this field is only present
		in this response if you set the access_type parameter
		to offline in the initial request to Google's
		authorization server.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>
Reviewed-by: Peter Tisovčík <xtisov00@stud.fit.vutbr.cz>