AbstractTransaction: extract abstract class
The Transaction now works more as an interface. This allows to
provide a more generic method impl() that returns the particular
TransactionImpl instance. The AbstractTransaction holds a single
instance of a selected TransactionImpl.

However, in case of a more complex transaction requirements (e.g.
chained transactions on multiple different resources), the Transaction
interface can be implemented in a different way to support it.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>