ThreadLocalTransaction: remake it to be just a wrapper
There is no reason to have the thread-local-specific transaction implementations.
The location where transactions are stored during runtime is independent of the
transaction type.

Thus, the original ThreadLocalTransaction is removed. Transactions should inherit
from AbstractTransaction or Transaction. Instead, there is now the class
ThreadLocalTransactionWrapper. The wrapper owns a pointer to the target transaction
and delegates calls to it.

The purpose of the ThreadLocalTransactionManager implementation is now to create
the appropriate ThreadLocalTransactionWrapper in the current thread and assign
the created Transaction with it. The ThreadLocalTransactionManager is not a
subject of functional changes in this commit.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>