ThreadLocalTransactionManager: separate transaction creation
Transactions are to be created by TransactionFactory implementations.
A TransactionManager manages a running transaction in some way and
allows its users to access it.

PocoTransactionManager and NoTransactionManager have been modified
to inherit TransactionFactory instead. Thus they are renamed and
considered to be factories only.

There are now only two TransactionManagers:

* NullTransactionManager
* ThreadLocalTransactionManager

The ThreadLocalTransactionManager is now a real manager that is to
be instantiated and that uses some TransactionFactory instance to
create transactions. The created transactions are managed using the
thread-local strategy.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>