dao: base: move transaction API into base
The transaction API is now located in base in directory transaction/.
Thus, we have to change all paths in the server to match this change.
The transaction API is built as a separate shared library, thus we
have to link it explicitly.

No functional change is done in this commit.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>