GatewayRPC: support asynchronous calls
GatewayRPC interface methods are extended by parameter std::function<void(GatewayRPCResult::Ptr)>
referred as ResultCall. It allows caller to pass function with reaction on GatewayRPCResult. It is
primary intended for asynchronous rpc, so this method can be stored and invoked when the result
arrives. In the LegacyGatewayRPC and FakeGatewayRPC it is invoked before returning from the RPC
methods.

All implementations of the GatewayRPC MUST execute ResultCall with final
result (NOT_CONNECTED, TIMEOUT, SUCCESS, FAILED) or aplication can
unexpectedly hang.

Example:

  Poco::Event event;
  GatewayRPCResult::Ptr localResult;

  rpc->pingGateway([&]GatewayRPCResult::Ptr result)
  	{
  		localResult = result;
  		event.set();
  	},
  	gateway
  );

  event.wait();

  Do something with localResult->status() ...

Signed-off-by: Jozef Halaj <xhalaj03@stud.fit.vutbr.cz>
Reviewed-by: Jan Viktorin <iviktorin@fit.vutbr.cz>