GoogleAuthProvider: FacebookAuthProvider: extract users locale
Make AuthProvider implementations to support extraction of user locale.
A locale should be obtained from the API server we use for
authenticating users.

We have to request also the locale scope throughout the authentification
process. The AuthResult holds the locale for upper layers to process it.

GoogleAuthProvider has the locale scope in the request as default.

Signed-off-by: Peter Kubov <xkubov06@stud.fit.vutbr.cz>
Reviewed-by: Jan Viktorin <iviktorin@fit.vutbr.cz>