SqitchLine: new abstract class
The SqitchLine represents a single line in a Sqitch plan. Each line
can have an optional note. Also, the line number is preserved for
debugging purposes. The line numbers allows to reconstruct also the
empty lines in a Sqitch plan.

The SqitchLine is sorted by the line number as default.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>