SqitchChange: new class
The SqitchChange represents a line containing a change of the database
schema. A change in Sqitch usually consists of 3 scripts: deploy, verify
and revert. Either deploy or revert is applicated to the database in
case when the change is in a given Sqitch plan. The plan can contain
e.g.:

(1) add-table
(2) +add-table
(3) -add-table

The lines (1) and (2) are equivalent, the deploy operation is executed.
The line (3) executes the revert operation of the particular change.

Every change can optionally have a list of dependencies and conflicts.
If there is an unmet dependency or a conflict is preceding this change
in the plan, such plan should fail to execute.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>