SqitchTag: new class
The class represent a tag line in a sqitch plan. A tag is
defined by name, timestamp, author and author's email.

A change (or other line) belongs to a tag (a release version)
if it precedes the tag and there is no following change of the
same.

(1)  @v3
(2)  add-table
(3)  @v2
(4)  -add-table
(5)  @v1
(6)  add-table

Here:

* add-table@v1 refers to the line (6),
* add-table@v2 refers to the line (4),
* add-table@v3 refers to the line (2),
* add-table refers to the line (2).

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>