SqitchPlan: new class
The SqitchPlan represents a sqitch plan consisting of multiple lines
of types: change, tag, pragma, note. Those for line types are
represented by SqitchChange, SqitchTag, SqitchPragma and SqitchLine.

The SqitchPlan allows to lookup for a specific line by SqitchRef or
for a tag. The lookup is performed in the reversed direction (from the
end of the plan) to resolve tags properly.

The pragmas "project", "syntax_version" and "uri" are parsed. Other
pragmas are ignored.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>