SqitchRef: new class
The SqitchRef represents a reference to a line in a Sqitch plan.
The reference can have one of the following formats:

  <name>
  <project>:<name>
  <name>@<tag>
  <project>:<name>@<tag>

Format

  <project>@<tag>

is currently NOT supported.

There is also a pitfall, without other context information, it
can never resolve the following expression to true:

  add-table@v1.0 == add-table

which can be true, if there is no newer version of the add-table
change then the @v1.0. The compare mtehods SqitchRef::equals() and
SqitchRef::lessThen() would behave in a wrong way in this case.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>