Switch to PostgreSQL/ODBC exclusively
The server is now configured to run against a testing database as
default. The testing database can be started via db/script/postgresql.py.
It prints all connection parameters into the standard output. It is
necessary to set at least database.port because the script chooses a
port at random. Example:

  $ beeeon-server -c conf/xmlui-startup.ini -Ddatabase.port=41525

Also, name of the database schema has changed to beeeon. The schema/
directory is not used anymore.

There is a necessary database ABI change in PocoSQLSensorHistoryDao.
The query sensors_history.fetch.huge.interval does not accept the
initial 2 arguments anymore. There were used only due to lack of
argument reordering support in the previous SQL query generation.
In PostgreSQL, we can handle this issue easily.

The PocoSqitchInitializer is created in the testing mode (ui.debug=yes).
This leads to an automatic initialization of the testing database
with no dependency on the Sqitch program. In non-testing mode, the
PocoSqitchInitializer is not used to avoid touching an already existing
database schema and contents.

The INI and SQL files generated by tools/xsl are not installed anymore.
They are not necessary and considered as obsolete.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>