tools: implement starting script intended for testing
The script starts an instance of PostgreSQL engine and then starts
a properly configured beeeon-server.

It is possible to configure the beeeon-server via command line arguments
(passed unchanged) and via environment variables. For most use-cases,
the only chnage is to pass -Dui.debug=yes while starting from the
main directory:

  $ tools/server-testing-start.sh -Dui.debug=yes

The environment variable TIMEOUT allows to start the server to work
and kill it automatically after the specified number of seconds.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>