SqitchInitializer: new class
SqitchInitializer implements the unsafeDeploy() method. The
unsafeDeploy() applies scripts in order as specified in the
preset sqitch plan. It however does not perform all checking
operations as the original sqitch program does. Also, it does
not resolve versions. The purpose of the SqitchInitializer is
to simplify initialization of database for testing.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>