PocoSqitchInitializer: new class
Integrate SqitchInitializer with the PocoDaoManager. The PocoDaoManager
initializes database via any number of PocoDBInitializer instances.

The PocoSqitchInitializer performs unsafeDeploy() on the initialization
request from the PocoDaoManager. Thus, there is no check of Sqitch
metadata and no revert performed on failure. It executes the given plan,
i.e. applies the steps: deploy and verify or revert.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>