UIMockInit: replace by t/testing-data.sql
The UIMockInit class is the last of the older UI-server reimplemention
part. This code was used for database content initialization for
testing. It was however hardcoded inside the application and its
execution, especially while inserting the module values, was very
slow.

With the provided t/testing-data.sql script, it is now clearer that
the testing data are related to tests in the t/ directory and due
to the proper SQL statements, its execution is very fast.

The t/testing-data.sql is executed only when ui.debug=yes.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>