PocoSQLUserDao: utilize LocaleManager to fetch user locale
The parseIfIDNotNull has been duplicated, first works in the old
way (no locale parsing) and the second use the given LocaleManager
to parse also the fetched locale. The reason is that we sometimes
do not need to fetch the locale and the particular queries do
not return it anyway. This is the case of fetching LegacyGateway
and its owner.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>
Reviewed-by: Peter Tisovčík <xtisov00@stud.fit.vutbr.cz>