GatewayCommunicator: new class
The GatewayCommunicator is responsible for communication with gateways
connected to the server. It maintains a table of gateway connections.

Each connection is managed by the Poco::Net::SocketReactor that is
running in a separate thread. The reactor notifies the system about
available readable connections.

Readable connections are queued and worker threads from the associated
thread pool handle these connections.

When a GWMessage is read from a gateway connection, GWMessageHandler is
used to handle it.

Signed-off-by: Jozef Halaj <xhalaj03@stud.fit.vutbr.cz>
Reviewed-by: Jan Viktorin <iviktorin@fit.vutbr.cz>