GWResponseExpectedQueue: new class
The GWResponseExpectedQueue provides the ability to detect missing
gateway responses to sent requests.

Each request sent to the gateway is registered to this queue by the
RPCForwarder and unregistered after arriving. In case of the missing
response, the timeout response is forwarded by the RPCForwarder
(there is cross reference) and gateway connection is removed from
the GatewayCommunicator if it is inactive.

Signed-off-by: Jozef Halaj <xhalaj03@stud.fit.vutbr.cz>
Acked-by: Jan Viktorin <iviktorin@fit.vutbr.cz>