RoleAccessPolicy: rework to use assure(Action) pattern
The assureInvite has been designed to accept also the "as" argument
telling the access level of the user we are about to invite. This
would allow to the scheme where e.g.:

* admin can invite and granting any access level
* user can invite and granting only guest access level.

Such functionaliy is however useless and finally, it was not implemented.
Thus, the new approach discards the "as" argument in behalf of the
API unification.

Actions ACTION_USER_INVITE and ACTION_USER_GET can be invoked
in the gateway context while ACTION_USER_GET, ACTION_USER_REMOVE
and ACTION_USER_UPDATE can be invoked in the role context.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>
Reviewed-by: Peter Tisovčík <xtisov00@stud.fit.vutbr.cz>