LocationAccessPolicy: rework to use assure(Action) pattern
The ACTION_USER_CREATE and ACTION_USER_GET can be invoked in
the context of a gateway.

The ACTION_USER_GET, ACTION_USER_UPDATE and ACTION_USER_REMOVE
can be invoked in the context of a location.

Thus, we have two methods implementing the actions based
on the given arguments.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>
Reviewed-by: Peter Tisovčík <xtisov00@stud.fit.vutbr.cz>