DeviceUnpairWork: rewrite to by fully asynchronous
The DeviceUnpairWork unpairs a remote device. It calls the GatewayRPC and expects
an asynchronous behaviour. I.e. it must call the ResultCall (lambda) in a different
thread (to avoid mutex deadlocks).

The implementation does not poll database anymore. It however performs a single
database access to assure that the device that is being unpaired still exists
in database. This may prevent some never-ending work under few unexpected
circumstances.

The ResultCall can be called repeatedly with new status information. The result
status is saved inside the DeviceUnpairWork (inside the Work instance). However,
no persistence is performed in this step. There is currently no facility to
handle it. (It might be probably added into the ExtendedWorkExecutor later).

When the result status is changed, the DeviceUnpairWork can be woken up.
In case of PENDING and ACCEPTED, no wakeup is done. Just, the result status
is updated inside the DeviceUnpairWork.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>
Reviewed-by: Jozef Halaj <xhalaj03@stud.fit.vutbr.cz>