TypesSAXHandler: do not handle tags <from>, <to> and <transform>
No transformation is to be defined at this level. All data must be
transformed before entering the system. There might be many differnt
transformations of a same type from various values. Thus, it is not
suitable to cover such feature by a single tag on the server.

E.g. the gateway must know its devices and their specifics. The server
does not know this information. The gateway thus performs all
transformations transparently.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>