Sensor: new class
The class should cover the upcoming changes in the devices.xml and dynamic
sensors definition. Previously, the ModuleInfo was used only which is not
enough as it does not contain dynamic information like the current value
or user-defined name.

In the future, the Sensor class would represent a persistable entity.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>