dao: utilize executeSelect() when checking the number of results
The return value from Poco::Data::Statement::execute() does not seem to correlate
with the number of results from SELECT statements. Sometimes, a strange behaviour
was seen where the execute() returns 1 for a SELECT that provides no rows.

Thus, change the way of executing of SELECTs in places where the number of rows
is being checked. Now, the number of rows is taken from the Poco::Data::RecordSet
and should be the value we want.

This change may introduce some incompatible behaviour because now the result
check may have a slightly different semantics.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>