FCMToken: new class
Provide support to Firebase Cloud Messaging (FCM) tokens which are used in server
authorization to send user notifications. Every token is registered
by user with request to BeeeOn server. Every registered token is saved
in database with user ID so that it is possible to notify user about
events which happened or are happening and require users attention.

Signed-off-by: Peter Kubov <xkubov06@stud.fit.vutbr.cz>
Reviewed-by: Jan Viktorin <iviktorin@fit.vutbr.cz>