base: install: do not install test-suite-server and related
The test-suite-server binary depends on the libBeeeOnBaseTest.so library.
However, it is not clear, why the library is not installed during the
install/packinging process. When installing the resulted RPM, the
installer complains about unmet dependencies:

error: Failed dependencies:
	libBeeeOnBaseTest.so()(64bit) is needed by beeeon-server-2017.9.1-1.x86_64

This is to be solved later. We do not need the test suite to be included
in installation package for now.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>