startup: restructuralize startup principle
The startup INI files are restructuralized into another approach. The original
startup was done just by choosing the right configuration file. The configuration
file could start up the particular server component only (e.g. REST-UI only).
This is useful for testing, however, it is not very helpful to distinguish between
testing and production startup.

The new approach divides the startup into two configurations:

* testing-startup.ini - testing configuration useful for semi/full automatic
  testing; it is intended to run the integration tests and Jenkins tests

* server-startup.ini - production configuration used to run the real system

The testing-startup.ini is likely to be tuned to start only the certain
component. This is done by giving -Dapplication.di.runner=<component>. The
available components are:

* all - default, used to run full server
* xmlui - XML-UI API only (a fake gws is used)
* restui*) - REST-UI API only (a fake gws is used)
* gws - GWS only (with no user frontend)

  *) the "restui" is renamed from "ui" in this patch.

The tools/server-testing-start.sh is modified accordingly. The variable TARGET
now specifies the DI runner, thus instead of passing TARGET=restui-startup,
the new approach accepts TARGET=restui. The script automatically uses the
testing-startup.ini (can be changed but it is not very useful). Thus, the
testing server is started e.g. as:

  $ TARGET=restui tools/server-testing-start.sh

The shared configuration file conf/config.d/server.ini is dropped as it is
not needed anymore. The testing and production startup configurations now contains
the specific configuration entries.

It is assumed that the database is not initialized on startup when using
the production configuration. The production configuration starts the services
at ports 8010, 8011 and 8012 to avoid collisions with testing deployments.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>