startup: rename option ui.debug to testing.enable
The configuration option ui.debug has been renamed to testing.enable.
The new name describes the purpose more accurately. The option is used
to enable database initialization which is used only in the testing
mode and which is not bound to the UI only. Also, the permit auth provider
is enabled in this case.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>