Sanitize: base64 accepts also URL variant
The Base64 for URL contains '-' and '_' characters instead of '+' and '/'.
The Sanitize::base64 now accept both variants (but not at the same time).

The RegularExpression did not pass tests when using a single regex
in form ([+a-zA-Z0-9/]+|[-a-zA-Z0-9_]+)={0,2} which should be more
effective. Thus, the implementation tries first one expression and
then the other with the URL-friendly modification.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>