main: call configure_file to avoid passing GIT_ID
Passing of GIT_ID via compiler definitions (cmake: add_definitions) leads
to a rebuild with each commit. However, a commit usually changes only a
small portion of the project and thus a full rebuilt is unnecessary.

The GIT_ID is currently used only in main.cpp. Thus, we reconfigure only
this file. Later on, we can create some kind of config.h.in strategy.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>