t: compensate winter timeshift
Due to the winter time shift, some integration tests start failing. The reason
is that data for the last 1 hour are missing (the reason for the tests to fail)
and data for the moment of the shift are duplicated. Demonstration:

 SELECT NOW() - (i * interval '1 hour')
     FROM generate_series(1, 1000) AS i;

 ...
 2017-10-29 05:34:59.02674+01
 2017-10-29 04:34:59.02674+01
 2017-10-29 03:34:59.02674+01
 2017-10-29 02:34:59.02674+01
 2017-10-29 02:34:59.02674+02
 2017-10-29 01:34:59.02674+02
 2017-10-29 00:34:59.02674+02
 2017-10-28 23:34:59.02674+02
 ...

The result is:

xmlui:
 not ok 1 - test1_getlogs_intemp_1hour

restui:
 not ok 11 - testb_history_room_temperature_1hour

where data for the "last hour" do not exist.

We fix that simply by simplify the series generation. Previously, data have been generated
from some point in the past (e.g. NOW() - 2 weeks) to the present. Now, we are generating
in the reverse order (i.e. from NOW() to the past) iteratively. Thus, the newest data are
always generated and available.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>