tester: support Google OAuth 2.0
The tester is now able to login via the Google OAuth 2.0. It requires that
server allows to login via redirect URI pointing to the tester's URL. As
default, this is <scheme>://<host>/tester.html. If it does not find any
provider offering redirect_uri in this format, then Google OAuth 2.0 is
ignored.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>