base: update to change Translator behaviour
Adapt the IcuTranslator to meet the new semantics of the Translator.
No, the formatImpl() should return an empty string in case of a failed
translation. The Translator's logic would take care of the rest.

Thus, the IcuTranslator is now derived from Translator directly, the
NoTranslator is avoided from the inheritance as it is no longer need
for the failed translations.

The tests are updated to use mostly formatSure() and only the test of
the missing keys are using the format() method with a default string.

All other code is updated to use formatSure() or format() accordingly.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>