db: update to have gateway with timezone
PocoSQLGatewayDao reflects this change accordingly, the ABI calls are fixed.
A TimeZoneProvider instance must be injected into the PocoSQLGatewayDao.

PocoSQLRoleInGatewayDao has to be modified as it calls PocoSQLGatewayDao::parseMany
which now requires a TimeZoneProvider instance.

Note that a global symbol timezone exists in time.h and it can be easily
confused with a "timeZone" variable - thus we are using "zone" instead.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>