CORSFilter: add Vary: Origin header to make caches happy
Caching on the way between clients and the server may decide to cache
responses. Caching is performed based on the requests' URIs, however,
when responses differ for same URI because of headers, the caching
would break the system and it can also expose data to wrong client.

The Vary header helps to define headers that should be part of the
caching key together with the URI. The CORSFilter can reply different
responses based on the given value of Origin.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>