auth: facebook auth is separated for web-app and android
There will be multiple FacebookAuthProvider instances, each for a
different purpose. That is because OAuth 2.0 works in a different
way for web applications and for smart phone applications.

Thus, the instance facebookAuthProvider is renamed to facebookWebAuth
and another similar instance named facebookAndroidAuth is added.
The instances IDs differ for unique referencing from clients:
facebook-webapp and facebook-android.

Signed-off-by: Jan Viktorin <iviktorin@fit.vutbr.cz>